ad authentication user pool

+++ src/services/spaces/DeleteSpaces.ts
@@ -1,11 +1,18 @@
 import { DeleteItemCommand, DynamoDBClient } from "@aws-sdk/client-dynamodb";
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
+import { hasAdminGroup } from "../shared/Utils";
 
 
 
 export async function deleteSpaces(event: APIGatewayProxyEvent, ddbClient: DynamoDBClient): Promise<APIGatewayProxyResult> {
 
     if(event.queryStringParameters && ('id' in event.queryStringParameters)) {
+        if (!hasAdminGroup(event)) {
+            return {
+                statusCode: 401,
+                body: JSON.stringify(`Not authorized!`)
+            }
+        }
         const spaceId = event.queryStringParameters['id'];
 
         await ddbClient.send(new DeleteItemCommand({
